requirements.txt
djangorestframework>=3.15,<4.0
django>=4.2,<5.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
azure-storage-blob>=12.19.0,<13.0
azure-identity>=1.15.0,<2.0
requests>=2.31.0,<3.0


